requirements.txt
fastmcp
pydantic
pydantic-settings
httpx
uvicorn
gspread
google-auth
